Create tool for rough plots
quick-plot.py uses xarray to read in one of the stream files from the two
one-month test runs for the JRA-forced high-res run with BGC turned on. Some
assumptions on the location of the data are hard-coded (it may be useful to
relax those assumptions and turn this into a more flexible tool at a later
date).

environment.yaml provides a conda environment that supports this tool.